move randomness tests to nightly test
